[refactor] Refactor items and db service.
