+++ news_crawler/items/__init__.py
@@ -1,0 +1,2 @@
+from .base_article import BaseArticle
+from .vnexpress_article import VnExpressArticle
