+++ news_crawler/pipelines/comment.py
