+++ scrapy.cfg
@@ -4,8 +4,8 @@
 # https://scrapyd.readthedocs.io/en/latest/deploy.html
 
 [settings]
-default = vnexpress.settings
+default = news_crawler.settings
 
 [deploy]
 #url = http://localhost:6800/
-project = vnexpress
+project = news_crawler
